+++ cypress/e2e/textBox.cy.js
@@ -5,12 +5,10 @@ import variables from "../fixtures/variables.json";
 
 describe('Textbox testing', ( )=> {
     const homePage =new HomePage();
-   
-    beforeEach(()=> {
-        homePage.navigateHomePage(variables.URL.site);
-       
-    })
 
+    beforeEach(()=> {
+        homePage.navigateHomePage(variables.URL.site);  
+    })
 
    it('fill textbox', function () {
     homePage
@@ -20,15 +18,11 @@ describe('Textbox testing', ( )=> {
         .fillUserEmail(variables.formData.email)
         .fillCurrentAddress(variables.formData.currentAddress)
         .fillPermanentAddress(variables.formData.permanentAddress)
-        .clickSubmitForm();
-        // .getNameUser()
-        // .should("be.visible")
-        // .and("have.text", variables.formData.firstName);
-
-  
+        .clickSubmitForm()
+        .getOutput()
+        .should("contain", `Name:${variables.formData.firstName}Email:${variables.formData.email}Current Address :${variables.formData.currentAddress}`);
+    
    });
   
-
-
   });
